rename docs folder so asset does not remove it
