+++ packages/scripts/src/helpers/doc-builder/index.ts
