+++ packages/scripts/src/helpers/doc-builder/sidebar.ts
